fix: apply effect override fails
move back to hook :(
